Start work on new workflow and Typer cli interface
Very much WIP
