+++ perdoo/console.py
@@ -13,15 +13,6 @@ LOGGER = logging.getLogger(__name__)
 CONSOLE = Console(
     theme=Theme(
         {
-            "prompt": "green",
-            "prompt.border": "dim green",
-            "prompt.choices": "white",
-            "prompt.default": "dim white",
-            "title": "magenta",
-            "title.border": "dim magenta",
-            "subtitle": "blue",
-            "subtitle.border": "dim blue",
-            "syntax.border": "dim cyan",
             "logging.level.debug": "dim white",
             "logging.level.info": "white",
             "logging.level.warning": "yellow",
